+++ packages/coreutils/src/index.ts
@@ -1,6 +1,6 @@
 import type { Program } from '@cyberspace/kernel'
 import { ls, cat, cp, mv, rm, mkdir, rmdir, touch, tree, chmod } from './fsu.js'
-import { sudo, make } from './sudo.js'
+import { sudo, sudoProgram, make } from './sudo.js'
 import { edit } from './edit.js'
 import { less } from './less.js'
 import { fortune } from './fortune.js'
@@ -20,3 +20,4 @@ export const coreutils: Record<string, Program> = {
 }
 
 export { strerror } from './util.js'
+export { sudoProgram }